+++ src/Components/Section_Elements/Columns.js
@@ -14,7 +14,7 @@ const StyledColumns = styled(motion.div)`
   padding-left: ${padding.outsideElements.double};
   padding-right: ${padding.outsideElements.double};
   grid-gap: max(
-    ${padding.outsideElements.double},
+    ${padding.outsideElements.single},
     ${padding.insideElements.single}
   );
   box-sizing: border-box;
