+++ src/Components/Section_Elements/SectionHeroCases/Highlights.js
@@ -17,7 +17,7 @@ const StyledHighlights = styled(motion.div)`
   padding-top: ${padding.outsideElements.double};
   padding-bottom: ${padding.outsideElements.double};
   grid-gap: max(
-    ${padding.outsideElements.double},
+    ${padding.outsideElements.single},
     ${padding.insideElements.single}
   );
   box-sizing: border-box;
