+++ src/Components/Section_Elements/UNUSED_Overview.js
@@ -16,7 +16,7 @@ const StyledOverviewWrapper = styled(motion.div)`
   padding-left: ${padding.outsideElements.double};
   padding-right: ${padding.outsideElements.double};
   grid-gap: max(
-    ${padding.outsideElements.double},
+    ${padding.outsideElements.single},
     ${padding.insideElements.single}
   );
   box-sizing: border-box;
